flask_app/requirements.txt
flask==2.3.3
flask-cors==4.0.0
openai==1.3.8
python-dotenv==1.0.0
pandas==2.1.3